requirements.txt
supabase>=2.0
python-dotenv>=1.0
httpx>=0.27
pydantic>=2.0
fastapi>=0.110
uvicorn[standard]>=0.27
jinja2>=3.1
